Replace && rename class CleaningCalculator by CleningTimeCalculator, create class ProductFactory, move methods for products in ProductFactory
